Fix Field.copy_field implementation preventing required custom field params
The old copy implementation was not only more complicated and less
efficient than it had to be, but since it called the constructor without
arguments, it basically enforced that custom fields cannot have required
parameters without reimplementing copy_field.

A test was added not only demonstrating custom fields, but checking this
will now work.